Add solution Lab7

+++ app/Http/Controllers/CrudUserController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Posts;
 use Hash;
 use Session;
 use App\Models\User;
@@ -105,9 +106,23 @@ class CrudUserController extends Controller
      */
     public function deleteUser(Request $request) {
         $user_id = $request->get('id');
-        $user = User::destroy($user_id);
 
-        return redirect("list")->withSuccess('You have signed-in');
+        $isDelete = false;
+        //Check existing post
+        $post = Posts::where('user_id', '=', $user_id)->first();
+
+        //Check existing favorite
+        $favorities = User::find($user_id)->favorities;
+
+        if (empty ($post) && $favorities->isEmpty()) {
+            $isDelete = true;
+        }
+
+        if ($isDelete) {
+            $user = User::destroy($user_id);
+        }
+
+        return redirect("list")->withSuccess('Delete successful');
     }
 
     /**
@@ -176,7 +191,7 @@ class CrudUserController extends Controller
     {
         if(Auth::check()){
             // $users = User::all();//Lay tat ca du lieu trong ban user
-            $users = User::paginate(2);
+            $users = User::paginate(10);
             return view('crud_user.list', ['users' => $users]);//->with('i',(request()->input('page',1)-1)*2);
         }
